camera become chaos when autorotating
The main demo has separated  allowsLandscape and take a picture
functions which will omit the chaos when both are invoking. When I set
the allowsLandscape = true and take a photo, the camera would rotate
automatically to cause the layer wrong and part of the camera view
covered beneath for me to have no way to take a photo.
My solution is to get rid of the Camera sourceType when trying to
autorotate. It works for me.
Hope to help someone. And I have learned a lot from author.