Fixes assetType in the "Inline" case
The "Inline" case was using pickerController.assetType to specify photos only; but because it had a custom DKImageGroupDataManager, .assetType didn't do anything. This moves the "photos only" specification to where it will work.

As an aside, this is kind of the "band-aid" fix: it shows people how to do it, but it still leaves the confusing possibility that someone will use assetType and it won't do anything. Any thoughts on whether I should implement it better, and if so, how?