Merge from develop (#519)
* Swift 4.2

* Bumping version to 4.0.1

* bugfix: 修复仅使用@1x的资源图的bug (#496)

Change-Id: I9f06ecc5da964c06e3150dd4aae524a285d02a4d

* Fixes assetType in the "Inline" case (#501)

The "Inline" case was using pickerController.assetType to specify photos only; but because it had a custom DKImageGroupDataManager, .assetType didn't do anything. This moves the "photos only" specification to where it will work.

As an aside, this is kind of the "band-aid" fix: it shows people how to do it, but it still leaves the confusing possibility that someone will use assetType and it won't do anything. Any thoughts on whether I should implement it better, and if so, how?

* Updated swift version in podspec (#510)

* Carthage (#516)

* Swift 4.2
* Carthage
* @objc

* cleanup

* pods

* carthage

* Pods (#517)

* cleanup
* carthage

* 4.0.3

* Updated.

* Updated.